Refactor xml generation responsibility
As expected, putting the xml-generation in a util type function will
be messy. Så I try and create an interface to use so that the domain
classes it self knows how to create xml them selves. The rootElement
is the only type that is supposed to be the xml-outputter
to an actual xml-file.