New interface to get xml-Element from domain classes + Added xml for Amount and LegalMonetaryTotal
The interface XmlElement will return a xml Node version of the
object. Notice that the name of the class will be the name of the
xml node. For this reason I added specific typing to variants of Amount.
In the domain there is a datatype called Amount and an element called
Amount. Subclasses of Amount will then just get a different xml-node-name.